+++ packages/iov-bcp-types/src/connection.ts
@@ -135,6 +135,8 @@ export interface BcpQueryTag {
 
 export interface BcpTxQuery {
   readonly id?: TransactionId;
+  /** send transaction to or from this address */
+  readonly sentFromOrTo?: Address;
   /** chain-specific key value pairs that encode a query */
   readonly tags?: ReadonlyArray<BcpQueryTag>;
   readonly height?: number;