+++ packages/iov-bns/types/index.d.ts
@@ -1,5 +1,5 @@
 export { bnsCodec } from "./bnscodec";
 export { bnsConnector } from "./bnsconnector";
 export { BnsConnection } from "./bnsconnection";
-export { bnsFromOrToTag, bnsNonceTag, bnsSwapQueryTags } from "./tags";
+export { bnsNonceTag, bnsSwapQueryTags } from "./tags";
 export { BnsBlockchainNft, BnsBlockchainsByChainIdQuery, BnsBlockchainsQuery, BnsUsernamesByChainAndAddressQuery, BnsUsernamesByOwnerAddressQuery, BnsUsernamesByUsernameQuery, BnsUsernamesQuery, BnsUsernameNft, BnsTx, AddAddressToUsernameTx, SetNameTx, RegisterBlockchainTx, RegisterUsernameTx, RemoveAddressFromUsernameTx, } from "./types";