+++ packages/iov-dpos/src/index.ts
@@ -1,4 +1,3 @@
 export { Derivation } from "./derivation";
 export { Parse } from "./parse";
 export { Serialization, TransactionSerializationOptions } from "./serialization";
-export { dposFromOrToTag, findDposAddress } from "./tags";